pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mnp-cdx"
version = "0.3.0"
description = "Production-oriented MNP analytics platform (MVP)"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "mnpCDX Main Architect" }
]
dependencies = [
  "duckdb>=1.0.0",
  "openpyxl>=3.1.0",
  "pandas>=2.0.0",
  "typer>=0.12.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.0.0",
  "streamlit>=1.30.0",
  "plotly>=5.0.0",
  "pyyaml>=6.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0"
]

[project.scripts]
mnp-cdx = "mnp_cdx.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
